Systray Optional Build.
Option to also build Neptune with systray.
Fixes issues related to systray in MacOs.
