+++ pkg/ui/sys.go
@@ -1,28 +1,26 @@
 //go:build systray
 // +build systray
 
-package neptune
+package ui
 
 import (
 	"github.com/getlantern/systray"
 	"github.com/m1ndo/Neptune/pkg/sdata"
 )
 
-type Sys struct {
-	*App
+// Register systray
+func (ui *UiApp) SystrayRun() {
+	systray.Run(ui.OnReady, nil)
 }
 
-func (App *Sys) Init() {
-	go App.AppRun()
-	systray.Run(App.onReady, nil)
-}
-
-func (App *Sys) onReady() {
+// onReady() For systray
+func (ui *UiApp) OnReady() {
 	systray.SetTemplateIcon(sdata.IcoRes.Content(), sdata.IcoRes.Content())
 	systray.SetIcon(sdata.IcoRes.Content())
 	systray.SetTitle("Neptune")
 	systray.SetTooltip("Neptune")
 	systray.AddSeparator()
+	mShow := systray.AddMenuItem("Show", "Show the main app")
 	mStart := systray.AddMenuItem("Start", "Start the soundkeys")
 	mPause := systray.AddMenuItem("Stop", "Stop the soundkeys")
 	mRand := systray.AddMenuItem("Rand", "Use a random soundkey")
@@ -30,13 +28,16 @@ func (App *Sys) onReady() {
 	go func() {
 		for {
 			select {
+			case <-mShow.ClickedCh:
+				ui.MainWindow.Show()
 			case <-mStart.ClickedCh:
-				App.AppRun()
+				ui.AppIn.AppRun()
 			case <-mPause.ClickedCh:
-				App.AppStop()
+				ui.AppIn.AppStop()
 			case <-mRand.ClickedCh:
-				App.AppRand()
+				ui.AppIn.AppRand()
 			case <-mQuitOrig.ClickedCh:
+				ui.MainWindow.Close()
 				systray.Quit()
 			}
 		}
